requirements.txt
baseconvert==1.0.0a4
Django==1.10.4
djangorestframework==3.5.3
flake8==3.2.1
flake8-import-order==0.11
gunicorn==19.6.0
mccabe==0.5.3
pycodestyle==2.2.0
pyflakes==1.3.0
